cast magma_int_t up to 64-bits using (long long); in dense
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@9393 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ control/connection_mgpu.cpp
@@ -34,7 +34,7 @@ magma_int_t magma_buildconnection_mgpu(
         magma_setdevice( d );
         cudaGetDeviceProperties( &prop, int(d) );
         if ( ! prop.unifiedAddressing ) {
-            printf( "device %ld doesn't support unified addressing\n", long(d) );
+            printf( "device %lld doesn't support unified addressing\n", (long long) d );
             magma_free_cpu( deviceid );
             return -1;
         }
@@ -48,30 +48,33 @@ magma_int_t magma_buildconnection_mgpu(
             gnode[cmplxid][lcgpunb] = d;
             deviceid[d] = -1;
         }
-        //printf("device %ld:\n", long(d) );
+        //printf("device %lld:\n", (long long) d );
 
         for( magma_int_t d2 = d+1; d2 < ngpu; ++d2 ) {
             // check for unified memory & enable peer memory access between all GPUs.
             magma_setdevice( d2 );
             cudaGetDeviceProperties( &prop, int(d2) );
             if ( ! prop.unifiedAddressing ) {
-                printf( "device %ld doesn't support unified addressing\n", long(d2) );
+                printf( "device %lld doesn't support unified addressing\n", (long long) d2 );
                 magma_free_cpu( deviceid );
                 return -1;
             }
 
             /* TODO err = */ cudaDeviceCanAccessPeer( &samecomplex, int(d), int(d2) );
 
-            //printf(" device %ld and device %ld have samecomplex = %ld\n", long(d), long(d2), long(samecomplex) );
+            //printf(" device %lld and device %lld have samecomplex = %lld\n",
+            //       (long long) d, (long long) d2, (long long) samecomplex );
             if (samecomplex == 1) {
                 // d and d2 are on the same complex so add them, note that d is already added
                 // so just enable the peer Access for d and enable+add d2.
                 // FOR d:
                 magma_setdevice( d );
                 err = cudaDeviceEnablePeerAccess( int(d2), 0 );
-                //printf("enabling devide %ld ==> %ld  error %ld\n", long(d), long(d2), long(err) );
+                //printf("enabling devide %lld ==> %lld  error %lld\n",
+                //       (long long) d, (long long) d2, (long long) err );
                 if ( err != cudaSuccess && err != cudaErrorPeerAccessAlreadyEnabled ) {
-                    printf( "device %ld cudaDeviceEnablePeerAccess error %ld\n", long(d2), long(err) );
+                    printf( "device %lld cudaDeviceEnablePeerAccess error %lld\n",
+                            (long long) d2, (long long) err );
                     magma_free_cpu( deviceid );
                     return -2;
                 }
@@ -79,17 +82,19 @@ magma_int_t magma_buildconnection_mgpu(
                 // FOR d2:
                 magma_setdevice( d2 );
                 err = cudaDeviceEnablePeerAccess( int(d), 0 );
-                //printf("enabling devide %ld ==> %ld  error %ld\n", long(d2), long(d), long(err) );
+                //printf("enabling devide %lld ==> %lld  error %lld\n",
+                //       (long long) d2, (long long) d, (long long) err );
                 if ((err == cudaSuccess) || (err == cudaErrorPeerAccessAlreadyEnabled)) {
                     if (deviceid[d2] == 0) {
-                        //printf("adding device %ld\n", long(d2) );
+                        //printf("adding device %lld\n", (long long) d2 );
                         gnode[cmplxid][MagmaMaxGPUs] = gnode[cmplxid][MagmaMaxGPUs]+1;
                         lcgpunb                      = gnode[cmplxid][MagmaMaxGPUs]-1;
                         gnode[cmplxid][lcgpunb] = d2;
                         deviceid[d2] = -1;
                     }
                 } else {
-                    printf( "device %ld cudaDeviceEnablePeerAccess error %ld\n", long(d), long(err) );
+                    printf( "device %lld cudaDeviceEnablePeerAccess error %lld\n",
+                            (long long) d, (long long) err );
                     magma_free_cpu( deviceid );
                     return -2;
                 }
